Changed ThreadID - uint32 => uint64. Updated NETWORK_VERSION - 6 => 7. New GUI is backward compatible with previous version.

+++ BrofilerCore/Core.cpp
@@ -392,8 +392,7 @@ OutputDataStream& operator<<(OutputDataStream& stream, const ScopeData& ob)
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 OutputDataStream& operator<<(OutputDataStream& stream, const ThreadDescription& description)
 {
-	//FIXME, TODO: Remove cast to uint32
-	return stream << (uint32)description.threadID.AsUInt64() << description.name;
+	return stream << description.threadID.AsUInt64() << description.name;
 }
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 OutputDataStream& operator<<(OutputDataStream& stream, const ThreadEntry* entry)
